Add Product Review
Model:
* [NEW] Review

Paths:
* [NEW] POST /product/:id/review -> review#create

Reviews allow a user to leave a review about the product.

Reviews are attached to a Product, and a product can have many
reviews. A user can create their review with this API.

In the Review model, the rating is restricted between 1 and 5,
inclusive.

Other validations are handled via the database layer.