Sorts for reviews
Models:
* [Updated] Reviews - New indexes

Adds indexes on the rating and created_at columns (with a primary
on the product_id column since ratings, for now, will be looked up
by product).

This enables the following sorts:
* product, rating (ascending and decending on rating)
* product, created_at (ascending and decending on created_at)

created_at DESC will be the default sort. If the created_at sort is
not enabled when using GET /review, it will apply a secondary sort
sort of created_at DESC to provide consistency with results returned.