Rank products by reviews
Models:
* [Updated] product - Add rank column and index

Paths:
* [Updated] GET /product - Sort by default on the rank of the product

Gems:
* [Updated] sidekiq - 6.5.1 (from soft locked 6.2.0)

Dev Environment:
* Docker compose updated with redis and sidekiq services

DevOps:
* Need Redis cluster
* Need sidekiq Rails instances
  * Use REDIS_URL env var for the Redis call path
* Add REDIS_URL env var to API server for Redis call path

As reviews are created, the product is ranked based upon the sum of the ratings vs the total possible sum of the ratings. This occurs as a background operation through sidekiq. The eventing mimics a change data capture (CDC) when the review was created and a consumer receives and handles. Ideally, this “decouples” the creation from the ranking, as the creation should not know about the ranking (and as the product evolves, this is a more standard practice to have multiple unknown consumers for a broadcasted event).

When GET /product is called, this returns an ordered list of products by rank then created_at.

Miscellaneous updates:
* Request specs can be annotated with `realistic_error_responses` to mimic what Rails will handle when exceptions occur in controller calls
* Test updates to break out expectations to their own specific test